refactor(Mocks): rework Mocks
- Add Episode+Mock.swift
- Add Season+Mock.swift
