+++ Sclerosis/Models/Episode.swift
@@ -10,5 +10,6 @@ import Foundation
 struct Episode {
     let id = UUID()
     let name: String
+    let showName: String
     let releaseDate: Date
 }
