+++ Sclerosis/Models/Mocks/Episode+Mock.swift
@@ -38,34 +38,6 @@ extension Episode {
     static let chernobylStayCalmEpisode = Episode(name: "Пожалуйста, оставайтесь спокойны", showName: "Чернобыль", releaseDate: makeDate(2019, 5, 13))
 }
 
-extension Season {
-    // MARK: - Individual Season Mocks
-    static let bbSeason1Mock = Season(episodes: [
-        Episode.bbPilotEpisode,
-        Episode.bbWalletOrLifeEpisode
-    ])
-    static let bbSeason2Mock = Season(episodes: [
-        Episode.bbSevenThirtySevenEpisode,
-        Episode.bbGrippyEpisode
-    ])
-    static let gotSeason1Mock = Season(episodes: [
-        Episode.gotWinterIsComingEpisode,
-        Episode.gotKingsroadEpisode
-    ])
-    static let chernobylSeason1Mock = Season(episodes: [
-        Episode.chernobylOneTwentyThreeFortyFiveEpisode,
-        Episode.chernobylStayCalmEpisode
-    ])
-
-    /// Набор всех моковых сезонов.
-    static let allMockSeasons: [Season] = [
-        .bbSeason1Mock,
-        .bbSeason2Mock,
-        .gotSeason1Mock,
-        .chernobylSeason1Mock
-    ]
-}
-
 extension Episode {
     /// Плоский список всех моковых эпизодов из `allMockSeasons`.
     static let MockEpisodes: [Episode] = Season.allMockSeasons.flatMap { $0.episodes }
